backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.15
django-cors-headers>=4.3
xhtml2pdf>=0.2.15
lxml>=4.9
requests>=2.31
